fix(scroll-tracking): decouple scroll depth tracking from autocapture
PROBLEM:
Scroll events were not firing when data-autocapture attribute was missing
or set to false. This affected customers who wanted scroll tracking but
not full autocapture functionality.

ROOT CAUSE:
ScrollDepth was tightly coupled to AutoCapture class:
- ScrollDepth instance was created inside AutoCapture constructor
- AutoCapture was only initialized when data-autocapture="true"
- If autocapture was disabled, ScrollDepth was never instantiated
- Result: No scroll event tracking even though it's independent functionality

This architectural coupling meant:
✅ Pageview events worked (independent of autocapture)
✅ Pageleave events worked (independent of autocapture)
❌ Scroll events failed (coupled to autocapture)

SOLUTION:
Moved ScrollDepth initialization from AutoCapture to UsermavenClient:
1. ScrollDepth now initializes in initializeBrowserFeatures() independently
2. Moved scroll event listeners from AutoCapture to dedicated method
3. Removed ScrollDepth dependency from AutoCapture class
4. Scroll tracking now works regardless of autocapture setting

CHANGES:
- client.ts:
  * Added scrollDepth property to UsermavenClient
  * Initialize ScrollDepth in initializeBrowserFeatures() before autocapture
  * Created initializeScrollDepthListeners() to handle scroll events
  * Moved isPageRefresh() logic from AutoCapture to client

- autocapture.ts:
  * Removed scrollDepth property and initialization
  * Removed scroll, visibilitychange, and popstate event handlers
  * Removed isPageRefresh() method (moved to client)
  * Removed scroll-related event handling from captureEvent()

IMPACT:
- Scroll depth tracking now works independently of autocapture
- Customers can disable autocapture but still get scroll events
- No breaking changes - existing functionality preserved
- Better separation of concerns and modularity

Fixes: SC-114646
